model_deploy/requirements.txt
numpy == 1.19.5
sklearn == 0.0
nltk == 3.6.2
django == 4.0.1
djangorestframework == 3.11
gunicorn == 20.1.0